backend/requirements.txt
aniso8601==7.0.0
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
Django==3.1.7
django-cors-headers==3.7.0
django-extensions==3.1.1
django-filter==2.4.0
django-graphql-auth==0.3.16
django-graphql-jwt==0.3.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
graphene==2.1.8
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
idna==2.10
numpy==1.23.0
openrouteservice==2.3.3
pandas==1.4.0
promise==2.3
psycopg2-binary==2.8.6
pydot==1.4.2
PyJWT==1.7.1
pyparsing==3.0.9
python-dateutil==2.8.1
python-dotenv==0.17.1
python-Levenshtein==0.12.2
pytz==2021.1
requests==2.25.1
requests-toolbelt==0.9.1
Rx==1.6.1
singledispatch==3.6.1
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
thefuzz==0.19.0
urllib3==1.26.4
whitenoise==6.2.0
pydot==1.4.2
gunicorn
iso3166
pydantic==1.9.0
python-Levenshtein
